load and cache Elasticsearch/SQL data types
The 'SYS TYPES' query is now executed right after connectivity check
part of SQLDriverConnect().
The result is cached on the connection (es_type array member, one
element for each type).
When a query is executed and the results processed, the column data type
sent by ES will be used to associate the right es_type element to the
record associated with the column.

This allows stripping the record structure from members that are only
used for data type characterization, thus saving memory, as well as
"transparency", in case new types are added or existing changed.
(Some change would still be needed in the driver, due to way ODBC
specifies "size" and "precision" of a column.)