single quote table type arg in SQLTables()
The spec allows for this argument to contain a list of table types to
filter by. The list of types can be given as quoted or unquoted
tokens. ES/SQL only suppots the quoted synthax, so do the quoting in the
driver.
Note: no sanity check done by the driver, though (i.e. mixed
quoted/unquoted or invalid characters). This isn't done for the other
arguments either.