b/f: uncouple NDEBUG from TEST_API
The goal is be able to run the tests in both cases, when the asserts are
in the code or out.
This is now achieved by only defining TEST_API macro, when it's not
already done.
The Release and MinSizeRel build configs will define it,
so the test functions won't be exported anymore (i.e. tests can't be
run).
Debug and RelWithDebInfo will not define it, so tests will be possible.