b/f: fix buffer indexing with registry values
The values from the registry are read into a static buffer, indexed for
each entry into the esodbc_dsn_attrs_st structure.
Make sure that the indexing is done only for those values in the
Registry that have a correspondence in the structure.

Also, use SQLWCHAR explicitely instead of the TCHAR (should be same
values for Unicode compilation, but might generate issues later.)