convert ES/BOOLEAN to C BIT
ES/BOOLEAN is used in results to catalog queries, like SYS TYPES (used
by SQLGetTypeInfo()). Since this is a non-ODBC standard type, some
applications (like Excel) have troubles converting it to the
coresponding standard integer required by the standard.

This commit changes the handling of ES/BOOLEAN, converting it to the BIT
type (used by ODBC for bools).

The driver still accepts the ES/BOOLEAN type from the application,
though, if this is ES types-aware.