fix CI test build if c/o dir contains plus char(s)
The CI checks out the projects in a directory named by concatenating the
project name, branch and event separated by a plus character.
This interferes with cmd.exe 'copy' command when launched as a CMake
in a custom target, which uses the plus sign to concatenate source files
to copy.