link libcurl statically
Add static linking for libcurl and make it the default.

This is useful since the client app might use libcurl and have it loaded
already. Since in this case no second version of the DLL will be loaded
(by default), the driver would have to use that instance, which might or
not have all the necessary features compiled in and behave or not as
expected.