fix: change timeout vs connection timeout setting
Applications set the login timeout, rather than the connection one.
Since every request is a new authenticated REST connection, the two
timeouts need to be treated as one: if one of the two attributes is set
to non-zero (=timeout enabled), prevent a subsequent call from disabling
it (assuming the application uses the "other" attribute).