fix SQLFetch with multi-row binding
This commit fixes a bug where,
- if an application was binding an array (column- or row-wise); and
- ES/SQL returns a cursor,
SQLFetch would either return no results (if the cursor was "empty"), or
the wrong number of results.

The reason for it was that the position within the application array to
write next was store along inside the resultset_st structure. This
structure was re-init'ed with every single new result received from
ES/SQL.

The fix was to migrate from a recusrive SQLFetch implementation (which
could have lead to high stack usage in some configurations) to an
iterative one, which no longer required saving the array position within
a resultset or statement structure (but rather remain function-local).

The applications that use SQLGetData() were not affected.

This commit also fixes a different bug (also affecting multi-row
binding) where the row status array pointer was read off the wrong
descriptor (ARD, instead of the correct IRD).