fix: reduce UJSON4C stack consumption
UJSON4C uses a stack scratchpad buffer for Unicode decoding. This buffer
is 128K by default and can lead quickly to stack overflows; easily
reproducible with 32bit builds.

The change changes the buffer size to 1 wchar_t size (below the minimum
serialized JSON object lenght), which triggers right away an allocation
of the buffer.
This is just a work-around, a better fix being allocating this buffer
along the private heap/allocator the library uses when deserializing.

Attempted fix for #61.