fix: add "fake" answers to unsupported catalog fn
Some catalog functions are required in order for the minimum driver
level to be achieved, so that apps can work with it. However, not all of
them have a current backing in ES/SQL. For these, a fake local empty
answer is provided.

With this change, some 2.x apps can now inquiry the columns count and
type (which 3.x apps would typically assume), which didn't work before.

The rest remains unchanged: the app will fetch no row and the call
succeeds.