instal. builder will now receive ZIP file as param
Change the builder to receive the driver .zip file as command line
argument, rathern than have it search driver's build directory for it.

This will allow building concurrently an installer for x86 and x64.

This is also the reason why now cleaning is no longer a dpendency for
the other targets.