fix: no param string rtrim when len is given
If the given string parameter lenght is not SQL_NTSL, it must contain
its lenght without the terminator -> don't trim the last character.

Also, for binary data, if the lenght pointer is NULL, try the
OCTET_LENGTH field, that should also contain the lenght (API
redundancy); if that's also non-positive, fallback to string lenght
calculation.