Fix: parameter length handling, error reporting, timestamp handling (#85)
* add/correct comments

- fix a typo
- add justification for non-portable _swnprintf() usage.

* fix: no param string rtrim when len is given

If the given string parameter lenght is not SQL_NTSL, it must contain
its lenght without the terminator -> don't trim the last character.

Also, for binary data, if the lenght pointer is NULL, try the
OCTET_LENGTH field, that should also contain the lenght (API
redundancy); if that's also non-positive, fallback to string lenght
calculation.

* fix: no err'ing on negative OCTET_LENGTH field

The field should never be negative (there's no case where it could be
SQL_NTS[L]). However, there are cases where it is to be ignored => log
a warning only.

* fix: enable binary-to-interval conversion

- in this case, the buffer of an interval structure is expected.
- add a u.test for it.

* fix: string-to-interval param conversion len fix

- the length of the string to be converted to an interval is to be read
from the OCTET_LENGTH_PTR field as a first choice, not OCTET_LENGHT
(that should also contain it, but that seems to be used primarily with
in-out/out params).

* revert workaround for timestamp as long

- Elasticsearch has switched back to providing the timestamp as ISO8601
  string.

* fix: ensure the errors propagate correctly on conn

In case the connection fails in early stages (i.e. types loading or
versions check), make sure the message received from the server is
reported to the user. This requires copying the diagnostic messages
between DBC and STMT in these cases.

* fix: using macro casting for diag access

* fix: add request mode and client-id to test req.

Add the ODBC request mode to the JSON request object. This will
trigger early (@connection testing in DSN editor) the license error, in
case of missing appropriate ES license.

(cherry picked from commit ed21cd8167333ea40086bf3c704537547b00086e)