fix: drop static buffer usage for cursor in req. (#88)
The cursor received from ES/SQL is being sent back to it when fetching
the next result set "page". When building the request, it's value was
converted from SQLWCHAR to mb-string in a static buffer, to find out
first the space required for it in the request.

Since the cursor value is a Base64 encoding string, the conversion can
simply be done by ASCII conversion, and thus the buffer is no longer
needed.

This fixes the case where the cursor value received from ES/SQL goes
over the buffer size (4KB), thus generating a failed query.