fix: init all record fields
Some of the string record fields have been left 0-ed, which is
problematic when their value is asked for. This commit inits them all to
the empty string, so that the transfer can take place.

(Their value would normally be provided by the data source, but this is
now not available with the current ES/SQL API.)