add support for timestamps in local timezone
With this commit the driver will be able to use local timestamps towards
the application and do the translation between that and ES/SQL (UTC)

A connection string parameter controlls if this behavior is enabled; it
is disabled by default.

The timezone is read from the TZ environment variable, or the system if
that isn't set.

The driver will now also provide the SQL format of date/timestamp when a
DATE/DATETIME value is converted to string.