Add support for Cloud ID (#125)
* add support for Cloud ID

This commit adds the support for a new connection string paramter,
CloudID.
This is to contain the string offered on Elastic's Cloud administration
page. It facilitates configuring the ODBC connectivity, by having the
driver decode&read from this string Elasticsearch's address and port, as
well as setting the security parameter to the highest value

The config GUI has now a new text field for it. Using this field
disables the input text fields for Hostname, Port and the SSL group.
Reciprocally, if the Hostname field has any text input, that will
disable the Cloud ID field.

* change CloudID tooltip message

- as per review suggestions