add configurable mode for floats conversion
This commit adds a new config paramter to the connection string that
instructs the driver to convert the floats into the default, scientific
or auto-mode notation; auto meaning scientific if the value is small or
large "enough".

It also uses scale received from Elasticsearch for the respective type.

Also, on the consistency checks the driver does on descriptors, the
precision and scale are now also those read from Elasticsearch (rather
than local values).