enable per-request time offset calculation
This commit will enable the calculation of the time offset (timezone
plus DST) for every request (done when serializing a statement). This
should allow correct operation of the driver also while DST turns into
efect.

The commit also switches back the calculation of the time offset to
using POSIX functions.

A warning is now logged if TZ environment variable is used and the DST
timezone acronym part is set since the Win CRT "assumes the United
States' rules", applied also when the timezone is not a US one.