add new DSN param to control PVA escaping
This commit adds a new connection string parameter to control the
escaping of the patern value arguments in catalog functions.

The PVA arguments can use '_' and '%' as special characters to build
patern matching values. However, some applications (likely do to some
more drivers misbehaving) use these chars as regular ones, which can
lead to the server returning more data than the app intended.

With the auto escaping (on by default), the driver will inspect the
arguments and will escape these special characters if not already done
by the application. (This is not standards compliant, that's why it's a
config option.)

The commit also adds more integration tests for the catalog functions