Apply DST (#142)
* fix: find if DST is in effect and apply offset

The system timezone offset doesn't account for daylight saving. This
commit adds the check if DST is in effect, calculates its offset and
adds it to the timezone offset for a total time offset.

* enable per-request time offset calculation

This commit will enable the calculation of the time offset (timezone
plus DST) for every request (done when serializing a statement). This
should allow correct operation of the driver also while DST turns into
efect.

The commit also switches back the calculation of the time offset to
using POSIX functions.

A warning is now logged if TZ environment variable is used and the DST
timezone acronym part is set since the Win CRT "assumes the United
States' rules", applied also when the timezone is not a US one.

* minor: remove testing decorator

- convert_init() no longer used by unit tests

* local-UTC time conversions done now with C lib fns

- convert timestamps using the CRT available functions, instead of the
incorrect use of start-time UTC-offset;
- add unit tests for the DST conversions

* minor headers and comments fixes

(cherry picked from commit 0cf97aa1a759245bb0c4dfbf67d05e00aa98bbd7)