fix: use IRD as precision src. instead of ARD
When handling values received from ES/SQL, use the IRD descriptor
instead of the ARD, when reading data characteristics (like the
precision).

Also, change the integration tests that were setting ARD fields - like
precision - on string conversions. (This is wrong since the strings
don't have a precision property, so the driver should not - and now will
not - take this property into account.)