Fix: consider interval's precision. Allow non-aligned period values as interval encoding (#148)
* add the precision value to the interval types

The interval types with a seconds component have the seconds
precision as a property. With this commit the driver will read the
Datetime(/Timestamp) scale and use that as precision for the interval
with seconds component data type property, attached to the .precision
field of the records.

* fix: use IRD as precision src. instead of ARD

When handling values received from ES/SQL, use the IRD descriptor
instead of the ARD, when reading data characteristics (like the
precision).

Also, change the integration tests that were setting ARD fields - like
precision - on string conversions. (This is wrong since the strings
don't have a precision property, so the driver should not - and now will
not - take this property into account.)

* fix: decouple ES period format from interval type

ES/SQL transmits interval values in a ISO8601 period value. An
interval-to-period 1:1 mapping is possible, however ES/SQL doesn't
adhere to this mapping. For instance:
- instead of `P1DT1H` for `INTERVAL '1 1' DAY TO HOUR`, ES/SQL will send
  `PT25H` (which in a 1:1 mapping would be `INTERVAL '25' HOUR` literal,
  equivalent, but different from the original);
- `INTERVAL '61:59' MINUTE TO SECOND` is encoded as `PT1H1M59S` instead
  of `PT61M59S`;
- `INTERVAL '61' SECOND` is transmitted as `PT1M1S` vs. `PT61S`.

Because of this, the driver will now detect the mapping misalignment,
convert any interval in the "day/to/second" range to seconds and
re-calculate the values for the corresponding interval members from
that.