fix: add pyodbc as instance member (#156)
Import and add the pyodbc module as instance member to the testing
class. This is a fix to the attempt to make pyodbc optional, when using
the integration framework for loading data into an ES instance only.

(cherry picked from commit 4440834aeec6bdc0d605f99d2a973909e8b17f10)