Add timing logging for the REST request
This commit adds small logging information on what are the timings
involved in getting a reply back from the server:
- the time to first response byte;
- the total time to retrieve the entire response. This should make it
easier to spot the queries that take a longer time to service.

The commmit also moves a set of logging statements on a lower level,
allowing to better troubleshoot performance issues by generating less
verbose logs and making them easier to pass around.