introduce CBOR format support for REST payloads
This commit adds basic support for CBOR encapsulation, as an alternative
to JSON.

The format to use is connection-specific and configured in the
connection string.

The introduction stops short of supporting encoding/decoding of the
parameters and result sets values: it will allow building a
non-parameterized query and decoding the response object (either with a
result-set or with an error), but without unpacking the values in the
result set of the latter.
Server version querying is also carried over CBOR.

All communication is done either over JSON or CBOR, although the driver
will carry on if it receives a JSON response to a CBOR request (or the
other way around). However, driver-generated "fake" responses to catalog
queries are always JSON-formatted (makes maintenance of text responses
easier).

The commit also enhances the support of the Elasticsearch-formatted
errors (both JSON and CBOR) in that the "error" parameter will be parsed
if of a map type (generally an ES/SQL error) or passed on as is, if of a
string type. The previous behavior was to abandon parsing if "error"
wasn't a map and present the entire error answer to the user; this
wouldn't work well with a CBOR object now.