fix: only fail time*/str conversion on truncation
The timestamp/time to string truncation needs to fail only if current
status code (for current column) indicates that truncation occured; so
status code has been added to the branching evaluation.

Previously, only the current diagnostic code was used in the decision
(and checked if it indicated truncation), which was incorrect since this
could have been set by the conversion of a previous column in same row.