fix length read from app ptr when param binding
This commit fixes the length handling of received wide-char strings as
parameters: the indicator-length pointer of the API call provides the
octet count, so the lenght needs to be divided by wide-char size.

Besides a couple of tests for the fix, the commit adds a new test on
null generation (both of non-NULL and NULL types).