SQLNumParams now counts parameter markers
This commit adds an simple counter of the parameter markers within an
attached SQL statement.

While a correct and complete implementation would require in our case
sending the statement to the server for parameter analysis, this simple
implementation should work for most cases where the application simply
wants to validate the number of user-provided paramters number against
the number of markers in the statement.