CBOR support for parameters
This commit adds the support for sending the paramters encapsulated as
CBOR.

The implementation makes use of the already available paramter conversion
functionality, which prepares the parameters as string values.
Consequently, except NULL values (and type), the values will be sent as
strings and have the server do a final conversion to the signaled SQL
type (which is actually also the case with JSON).