add CBOR integration testing
This commit adds a re-run of the integration tests, but carried over
CBOR.

The commit also enables parameter testing, but modifying the "protocol"
tests and adding an equality test with a parameter for each of them; the
paramter is the same value that the tests use to fetch from the server.

With this commit the application now queries the name of the cluster to
match the cataog name against (so that it can be run against nightly
build or a local compmilation easily).