fix size of DATE parameters
This commit fixes the handling of the columns size of the DATE type parameters.

The previous implementation always expected the size of an ISO8601
timestamp, since DATE and TIME vals area ultimately parsed as a
timestamp. However the column size was passed through unchanged. With
this commit, the size is read as passed in by the application and
adjusted as needed.