Fix handling of size with DATE parameters (#178)
* fix size of DATE parameters

This commit fixes the handling of the columns size of the DATE type parameters.

The previous implementation always expected the size of an ISO8601
timestamp, since DATE and TIME vals area ultimately parsed as a
timestamp. However the column size was passed through unchanged. With
this commit, the size is read as passed in by the application and
adjusted as needed.

* address PR review notes, fixing logging messages

- fix missing arguments to logging macros;
- improve phrasing.

* address PR notes: fix logging format specifier

- s/%d/%zu

(cherry picked from commit 24fb40e04a16c8bf08702aac71466e02804d933e)