Use default float representation for parameters (#179)
With this commit the float parameters are no longer sent in the
exponential format, but the default one. With it, the precision of the
sent floats is also increased and respecting the requirements given by
Elasticsearch for each of the float types.

(cherry picked from commit 92605c1b202f3e3e0f378bee00d2aaa74af68906)