SQLRowCount: always return row set size (#182)
With this commit SQLRowCount() will always return the size of current
row set, also in the case a cursor is present.

The previous implementation returned 0 when a cursor was present, since
the function can't evaluate the total result set size (without actually
fetching it all) and applications shouldn't rely on this result.

The driver - that internally also uses the function for SYS queries -
assumed that ES/SQL will never return a cursor for SYS commands. This
seems however not to be a valid assumption.