support of the new SHAPE data type
This commit adds support for the new SHAPE data type. Supporting it
requires acceptance of the data type -- the type is mapped as a text
object (like the GEO_* types) and the driver treats it as an opaque text
field

The commit also updates the set of SQL functions the driver reports as
supported by Elasticsearch (as data source).