fix: don't check logging path if logging disabled (#186)
This commit fixes an issue where the testing of a connection or saving
of the DSN failed when the given logging path was invalid (incorrect or
directory missing), even when logging was disabled.
The fix is to simply skip path validation if logging is disabled.

The commit also changes the tab index of the components of each tab
page, so that tab cycling follows the visual order of the elements.

(cherry picked from commit a6a9641555e4959a5a64cdf697c899fa183ca84c)

Cherry-pick conflicts:
- dsneditor/EsOdbcDsnEditor/DSNEditorForm.Designer.cs
resolved by keeping branch version (tab index fix not applied).