sign DLL files to be added into the MSI
So far only the MSI had been considered for signing. This commit adds
the logic to also sign the DLL files that are going to be packed into
the MSI.

The commit adds the code to find all the DLL files within the directory
unpackaged from the ZIP file produced by the build, then sign them with
same function used to sign the MSI file.