fix buffer underrun for wrong received buffer size (#199)
This commit fixes an underrun in the utility function copying a
wide-char string to the client application.

The impact is limited and it can only happen if the
received size of the client buffer is of the wrong size (<=0).

The commit also fixes a logging error happening if this function needs
to truncate the output. The macro for a format specifier with no
precision was provided instead of the needed one with precision. This
could potentially crash the client application in case the string to be
logged in not 0-terminated, logging is explicitely
enabled at an INFO (or higher) level and truncation needs to be applied

(cherry picked from commit ea2814b5609bf7f2b5b3a33767cb68a26283b264)

Merge conflicts:
- test/test_util.cc
fixed by removing test for utf8_to_wstr(), not present in 6.x.