Fix: init var that's otherwise accessed uninit'd (#207)
This commit initializes a variable ("format") that can be read
uninitialized. There are no negative manifestations of this issue.

The var carries the type of time format that has been detected when
parssing a string input value, which would subsequently be converted to
a TIME/DATE/TIMESTAMP data type. In case the source is a string
representing a TIMESTAMP and the destination a DATE, the representation
of the destination is obtained from the original input, by zero-ing the
time part (since this is how DATEs are represented in ES/SQL) and the
"format" var decides, in a disjunction with another variable, on this
action.

In case the source is of the type DATE and the destination also a DATE,
the "format" var would not be set, but its value read. In the case the
random value the var can take matches the condition, a zero'ing of an
already zero'd string would happen, with no other side-effect.

(cherry picked from commit d1392c76bb4ba6d7543086e4a0d98896ab6f80a3)