allow (re)setting the current catalog
This commit allows the application set the value of the current catalog
attribute, in case this value is the same with what the driver had
previously returned for this attribute.

This currently only works if the attribute had been retrieved before
setting, on the same connection; specifically, it won't work when this
would be set as a pre-connection attribute.