allow some statment attrs be used for a connection
Allow two statment attributes be used on a connection handle.

This is in support of some 2.x applications that can make use of them.
- SQL_ATTR_MAX_ROWS: the value is reset to 0 if different, which is the
same behavior the driver has when the attribute is set on a statement;
- SQL_ATTR_QUERY_TIMEOUT: the attribute is applied; any statement
belonging to the affected connection can subsequently override the
value, so the support is safe.