Fixes: possible buffer overrun on interval conversion. Catalog length handling (#217)
* fix overrun when converting a c-string to interval

When converting a C-string to a wide string, the conversion function
0-terminates the output. However, the allocated destination buffer
lacked the space for the terminator. This commit fixes that.

* fix catalog setting

The API function provides the byte count for the wide string name of the
catalog, not the character count, as so far implemented. This commit
fixes the lenght handling.