Params: ignore unbound parameters (#223)
Unbound parameters are allowed by the standard, although there's a
mechanism that allows the application to signal the driver which
parameters to ignore. There are appliacations however that set certain
attributes of a first parameter (despite the query containing no
markers), that however don't use this mechanism.

This commit will have the driver ignore parameters with attributes set
by the application, but not bound, as long as these unbound ones don't
show up before the contiguous list of bound ones.