Fix: SQLGetData: only update offset if C-string data copied out (#226)
* Fix: SQLGetData: only update offset if data copied

This commit fixes a defect manifesting when getting C-strings with
SQLGetData, in two steps: a first invocation with a 0-space buffer, to
fetch the lenght of data available, then a subsequent one, with
appropriate buffer size, to actually fetch the data.

The second call would so far fail with "no data available anymore",
since the offset of so-far-read-data was incorrectly updated, also in
the case when no data was actually copied. This commit fixes it, only
updating the offset if there's space in output buffer to copy.