Update type selection for string parameters (#218)
* proper cleans all builds

When invoking 'proper', the build script now invokes the build scripts
for both x86 and x64, if any of them are present.

* inc uninstallation timeout while integ-testing

wmi application used to uninstall driver by name during integration
testing can take over 1.5 minutes on slower VMs with many installed
applications.

* extend logging buffers for (CBOR) hex printers

When extended logging debugging define is enabled, the hex dumping of
binary buffers (used for CBOR replies logging) are now also dynamically
allocated and significantly larger (4k -> 4M).
This allows easier troubleshooting.

* change the ES type selection for text parameters

This commit changes the way a parameter's SQL type is resolved to an ES
type. So far, the size of the parameter was taking into account, to
select the ES type with the lowest "precision" that would accomodate it.

This worked well so far because IP type was having a precision of 0.
This precision for IP type had been updated in Elasticsearch (to maximum
IP length), which would lead to the driver tagging small length strings
as IP types.

With current change, all string type parameterss are considered as TEXT
types, leaving Elasticsearch/SQL to apply a better, context-informed
conversion.