Introduce new varchar limitation option (#229)
* introduce new varchar limitation option

This commit adds a new feature that allows the user to limit the size of
the string types (TEXT, KEYWORD & co.). This is needed for those
applications that won't accept the large maxes that ES does for the
fields of these types.

The change will manifest itself as:
- limitation of the COLUMN_SIZE, BUFFER_LENGHT and CHAR_OCTET_LENGTH in
the SQLColumns output;
- limitation of the "display size" attribute in the description of the
columns of the above mentioned types;
- truncation of the results to the configured limit for the columns of
the above mentioned types.

The commit also adds unit tests for the "faked" (i.e. driver-returned
results) catalog functions, removes stale DSN attribute definition (left
over from #215) and corrects the row index in a debug statement.

* add new integration testing running mode

Add new option to execute only the tests requiring no indexing of data
(currently server info-fetching tests and the protocol tests).

The commit also sets the autocommit as parameter to connection function,
instead of the subsequent operation.