Fix ES varchar types mappings to ODBC types
This commit fixes the mappings of the varchar types to the ODBC data
types:
- the KEYWORD, CONSTANT_KEYWORD and TEXT types are now SQL_WVARCHAR,
which allows certain applications to correctly read the UTF8 content.
- IP and the other string types (like GEO ones) remain to the previous
SQL_VARCHAR types.

The handling of double types has also been fixed to reflect the
correction in Elasticsearch/SQL that now maps DOUBLE and SCALED_FLOAT to
the same type (SQL_DOUBLE).