Fix debug printing of SQL(U)LEN variables
This fixes the debug printing of SQL(U)LEN variable types.
This type is defined as a long on WIN32, but a __int64 on WIN64.
The "%llu"/"%lld" specifier had be used from the start, before the need
for an x86 driver was considered and stayed like that mostly since.

The values logged by an x86 drivers contains the actual values of the
variable, but one needs to convert to hexa, remove the upper 4 byte and
convert back, which is cumbersome.

The values are now promoted to a (u)int64_t in the logging statements,
so printf will always log the correct value.

This change also updates some other few incorrect specifiers.