Fix varchar limitation when max-len-attr is set (#231)
* Fix varchar limitation when max-len-attr is set

This fixes the result rewriting of the SQLColumns due to varchar limit
setting in case the SQL_ATTR_MAX_LENGTH attribute is set. The attribute
changes the lenght indicator given by SQLGetData to the value of the
attribute (which a setting application expects).
Since the rewriting is serialized on the statment, simply reset this
attribute's value while doing the writing.

The types of the columns in the rewritten result are also corrected
(they were copied from ES/SQL output, which are wrong and need fixing).