add support for chunked strings
ES will split the large string values in chunks (of ~4KiB), even though
the entire set of chunks is contained within the answer (and thus
needless). This commit:
- introduces a new CBOR utility function that will fail to get string
references, if the value is chunked; this is mostly useful for fetching
object key names, which are part of the protocol and short;
- will now allow the cursor string to be chunked, in which case it'll be
allocated by the tinycbor library.