Log credentials used to set up a connection
This commit adds info-logging of the credentials (with redacted
password), just like with the rest of the connection properties.

The properties are also debug-logged when the connection string is
parsed, but those messages aren't logged into the per-connection log
file. (The user needs to configure the "alternative logging" to get
those, which is more convoluted.)