Log credentials used to set up a connection (#256)
This commit adds info-logging of the credentials (with redacted
password), just like with the rest of the connection properties.

The properties are also debug-logged when the connection string is
parsed, but those messages aren't logged into the per-connection log
file. (The user needs to configure the "alternative logging" to get
those, which is more convoluted.)

(cherry picked from commit 16c811cc9aa3210ee285b15c19d73036aa801422)