Testing convenience: build DSN out of URL
In case the '-p' parameter provides a URL to connect to a running
Elasticsearch instance, use that to construct a testing DSN, so another
(and possibly conflicting) '-c' paramter won't need to.

This also fixes a defect where trying to run the integration suite in
non-indexing mode fails due to missing meta-data.