Update CMake generation for Visual Studio 16 (#249)
* update CMake generation to Visual Studio 16

Update the how CMake generator specs to build with Visual Studio 2019
16.

CMake 3.14 or newer is required for it.

* add support for a multi-VS environment

This adds support for both VS 2019 and 2017 editions being installed.

The 'setup' will now take an optional parameter that dictates which of
the editions to set up. If none is provided 2019 and 2017 are looked up,
in that order.

CMake now always receives the generator to use (through '-G' param) and
target platform ('-A' param). The former is detected from the
environment variables (VSCMD_VER), the latter, as before, in order of
priority: cmd line param, env. var (VSCMD_ARG_TGT_ARCH) or registry key.

* Update build.bat

Co-authored-by: David Roberts <dave.roberts@elastic.co>

* Fix: force script end on VS misdetection

If setting a generator name fails, propagate the error and quite the
script.

Co-authored-by: David Roberts <dave.roberts@elastic.co>