Add catalog support
This adds support for multiple catalogs ("databases").
The client application can now set the catalog at connection level -
with SQLSetConnectAttr - to a pattern matching a local or remote cluster
name and ES/SQL will execute the query on the respective cluster/s.

A default catalog can also be set through the connection string, through
the "Catalog" DSN attribute.

In case the catalog is not set, or set to an empty or null string, the
local catalog will be chosen for execution.

The driver will simply add a "catalog" request attribute containing the
value received from the application, or skip adding the attribute if
this is empty.