Add unsigned_long ES type support
This adds support for the ES/SQL UNSIGNED_LONG type.

Values with this type can be received by the driver and forwarded to the
application as SQL_C_UBIGINT types. The driver will also accept
parameters of this type.

The driver-internal conversion to/from UNSIGNED_LONG is also supported.

The commit also adds logging of the name of the application loading the
driver.