Add support for API Key authentication (#313)
This adds support for authenticating using API Keys.

The DSN editor has been updated to take a new "API Key" field. This
field is enabled iif no Username and Password are provided.
The field is stored in the DSN string and then used in every request for
respective configured connection.